freebsd-{lib,share}-9.9999: fix die "failed to set sys symlink"

+++ sys-freebsd/freebsd-lib/freebsd-lib-9.9999.ebuild
@@ -177,7 +177,9 @@ src_prepare() {
 	fi
 
 	if ! is_crosscompile ; then
-		ln -s "/usr/src/sys" "${WORKDIR}/sys" || die "Couldn't make sys symlink!"
+		if [[ ! -e "${WORKDIR}/sys" ]]; then
+			ln -s "/usr/src/sys" "${WORKDIR}/sys" || die "Couldn't make sys symlink!"
+		fi
 	else
 		sed -i.bak -e "s:/usr/include:/usr/${CTARGET}/usr/include:g" \
 			"${S}/libc/rpc/Makefile.inc" \
